Refactor: move client auth into dedicated modules
- Extracted auth fetch calls to public/js/auth/api.mjs
- Moved auth UI logic (login/logout modal, authbox) into public/js/auth/ui.mjs
- Updated controllers.mjs to use makeAuthUI for session state
- Added tests/auth/helpers.mjs with loginAs/logout helpers
- Updated juices.put-liters.spec.mjs to use loginAs
- Verified juices.authz.spec.mjs covers 401/200 authz flows

Refactor: move client auth into dedicated modules

- Extracted auth fetch calls to public/js/auth/api.mjs
- Moved auth UI logic (login/logout modal, authbox) into public/js/auth/ui.mjs
- Updated controllers.mjs to use makeAuthUI for session state
- Added tests/auth/helpers.mjs with loginAs/logout helpers
- Updated juices.put-liters.spec.mjs to use loginAs
- Verified juices.authz.spec.mjs covers 401/200 authz flows
