+++ tests/juices.authz.spec.mjs
@@ -1,6 +1,7 @@
 // tests/juices.authz.spec.mjs
 import { describe, it, beforeEach, expect } from 'vitest'
 import { makeApi } from './helpers.mjs'
+import { loginAs } from './auth/helpers.mjs'
 
 describe('Authz on write routes', () => {
   let api
@@ -13,11 +14,7 @@ describe('Authz on write routes', () => {
   })
 
   it('PUT /api/juices/:id/liters → 200 when logged in', async () => {
-    // first-time set pin for user 1 (starts session)
-    await api.post('/api/auth/begin').send({ userId: 1 }).expect(200)
-    await api.post('/api/auth/set-pin')
-      .send({ userId: 1, pin: '1234', confirm: '1234' })
-      .expect(200)
+    await loginAs(api, 1)
 
     const res = await api.put('/api/juices/1/liters')
       .send({ liters: 5 })
