+++ tests/juices.put-liters.spec.mjs
@@ -1,6 +1,7 @@
 // tests/juices.put-liters.spec.mjs
 import { describe, it, beforeEach, expect } from 'vitest'
 import { makeApi } from './helpers.mjs'
+import { loginAs } from './auth/helpers.mjs'
 
 /**
  * PUT /api/juices/:id/liters
@@ -15,11 +16,7 @@ describe('PUT /api/juices/:id/liters', () => {
   let api
   beforeEach(async () => {
     api = await makeApi({ seed: true })
-    // Authenticate the agent once for all tests in this suite
-    await api.post('/api/auth/begin').send({ userId: 1 }).expect(200)
-    await api.post('/api/auth/set-pin')
-    .send({ userId: 1, pin: '1234', confirm: '1234' })
-    .expect(200)
+    await loginAs(api)
   })
 
   it('updates currentLiters, returns updated record, and bumps lastUpdated (list-only flow)', async () => {
